+++ browser/environment.js
@@ -7,9 +7,9 @@
  * Code distributed by Google as part of the polymer project is also
  * subject to an additional IP rights grant found at http://polymer.github.io/PATENTS.txt
  */
-import * as config    from './config.js';
+import * as config from './config.js';
 import * as reporters from './reporters.js';
-import * as util      from './util.js';
+import * as util from './util.js';
 
 /**
  * Loads all environment scripts ...synchronously ...after us.