+++ data/a11ySuite.js
@@ -26,8 +26,9 @@
           *
           * @param {String} fixtureId ID of the fixture element in the document to use
           * @param {Array?} ignoredRules Array of rules to ignore for this suite
+          * @param {Function?} beforeEach Function to be called before each test to ensure proper setup
           */
-        context.a11ySuite = function(fixtureId, ignoredRules) {
+        context.a11ySuite = function(fixtureId, ignoredRules, beforeEach) {
           // capture a reference to the fixture element early
           var fixtureElement = document.getElementById(fixtureId);
           if (!fixtureElement) {
@@ -50,6 +51,14 @@
           a11ySuite.eachTest = function() {
             // instantiate fixture
             fixtureElement.create();
+
+            // Make sure lazy-loaded dom is ready (eg <template is='dom-repeat'>)
+            Polymer.dom.flush();
+
+            // If we have a beforeEach function, call it
+            if (beforeEach) {
+              beforeEach();
+            }
 
             // run audit
             var auditResults = axs.Audit.run(axsConfig);